Break up project type/feature fixtures for manageability
