+++ platforms/core-configuration/declarative-dsl-provider/src/integTest/groovy/org/gradle/internal/declarativedsl/WorkingWithFilesIntegrationTest.groovy
@@ -18,6 +18,9 @@ package org.gradle.internal.declarativedsl
 
 import org.gradle.features.binding.BuildModel
 import org.gradle.features.binding.Definition
+import org.gradle.features.internal.builders.definitions.ProjectTypeDefinitionClassBuilder
+import org.gradle.features.internal.builders.settings.SettingsPluginClassBuilder
+import org.gradle.features.internal.builders.types.ProjectTypePluginClassBuilder
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.polyglot.PolyglotDslTest
 import org.gradle.integtests.fixtures.polyglot.PolyglotTestFixture
