+++ platforms/core-configuration/declarative-dsl-provider/src/integTest/groovy/org/gradle/internal/declarativedsl/settings/ProjectTypeModelDefaultsIntegrationTest.groovy
@@ -16,7 +16,6 @@
 
 package org.gradle.internal.declarativedsl.settings
 
-import groovy.test.NotYetImplemented
 import org.gradle.features.internal.ProjectTypeFixture
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.UnsupportedWithConfigurationCache
@@ -373,7 +372,6 @@ class ProjectTypeModelDefaultsIntegrationTest extends AbstractIntegrationSpec im
         outputContains("Foo(name = two, x = 2, y = 22)")
     }
 
-    @NotYetImplemented
     def "can configure build-level defaults in a settings plugin"() {
         given:
         withSettingsPluginThatConfiguresModelDefaults().prepareToExecute()
@@ -387,7 +385,7 @@ class ProjectTypeModelDefaultsIntegrationTest extends AbstractIntegrationSpec im
 
         then:
         outputContains("definition id = test")
-        outputContains("definition foo.bar = plugin")
+        outputContains("definition foo.bar = settings")
     }
 
     private static String[] expected(Map<String, String> expectations) {
