+++ platforms/core-configuration/declarative-dsl-provider/src/testFixtures/groovy/org/gradle/features/internal/builders/definitions/ProjectFeatureDefinitionThatUsesClassInjectedMethods.groovy
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.features.internal.builders.definitions
+
+/**
+ * A {@link ProjectFeatureDefinitionClassBuilder} that uses methods injected by the base class to set properties on the model.
+ */
+class ProjectFeatureDefinitionThatUsesClassInjectedMethods extends ProjectFeatureDefinitionClassBuilder {
+    @Override
+    String getBuildModelMapping() {
+        return """
+            model.getText().set(definition.getText());
+            model.getDir().set(getProjectFeatureLayout().getProjectDirectory().dir(definition.getText()));
+        """
+    }
+}
