+++ platforms/core-configuration/declarative-dsl-provider/src/testFixtures/groovy/org/gradle/features/internal/builders/definitions/ProjectTypeDefinitionThatRegistersANestedBindingLocationClassBuilder.groovy
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.features.internal.builders.definitions
+
+/**
+ * A {@link ProjectTypeDefinitionClassBuilder} for creating a project type definition that registers a nested binding location.
+ * The parent definition's build model is used in the build model mapping to set a property on the nested build model.
+ */
+class ProjectTypeDefinitionThatRegistersANestedBindingLocationClassBuilder extends ProjectTypeDefinitionClassBuilder {
+    @Override
+    String getBuildModelMapping() {
+        return """
+                context.registerBuildModel(definition.getFoo())
+                    .getBarProcessed().set(definition.getFoo().getBar().map(it -> it.toUpperCase()));
+            """
+    }
+}
