+++ platforms/core-configuration/declarative-dsl-provider/src/testFixtures/groovy/org/gradle/features/internal/builders/types/ProjectTypeThatUsesUnknownServicesClassBuilder.groovy
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.features.internal.builders.types
+
+
+import org.gradle.features.internal.builders.definitions.ProjectTypeDefinitionClassBuilder
+import org.gradle.features.registration.TaskRegistrar
+
+/**
+ * A {@link ProjectTypePluginClassBuilder} for generating a project type plugin class that uses an unknown service in its services interface.
+ */
+class ProjectTypeThatUsesUnknownServicesClassBuilder extends ProjectTypePluginClassBuilder {
+    ProjectTypeThatUsesUnknownServicesClassBuilder(ProjectTypeDefinitionClassBuilder definition) {
+        super(definition)
+    }
+
+    @Override
+    String getServicesInterface() {
+        return """
+            interface Services {
+                default ${TaskRegistrar.class.name} getTaskRegistrar() {
+                    return getUnknownService();
+                }
+
+                @javax.inject.Inject
+                UnknownService getUnknownService();
+            }
+
+            interface UnknownService extends ${TaskRegistrar.class.name} { }
+        """
+    }
+}
